Refactored the file system to take advantage of the buffer pool.  Now archiving semi-random data at 587,000 points per second.
git-tfs-id: [https://tfs.codeplex.com/tfs/TFS11]$/openHistorian/Main;C15441

+++ Source/Libraries/openHistorian.Core/Unmanaged/BinaryStream.cs
@@ -154,7 +154,7 @@ namespace openHistorian.Core.Unmanaged
         {
             //if (handle.IsAllocated)
             //    handle.Free();
-            long buffer;
+            IntPtr buffer;
             m_currentPosition = Position;
             m_stream.GetCurrentBlock(m_currentPosition, isWriting, out buffer, out m_firstIndex, out m_lastIndex, out m_currentIndex);
